+++ app/src/main/java/com/example/todoapp/ui/preview/EditTaskFormScreenPreview.kt
@@ -1,12 +1,12 @@
-package com.example.todoapp.view.preview
+package com.example.todoapp.ui.preview
 
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import com.example.todoapp.model.Task
+import com.example.todoapp.data.repository.model.Task
 import com.example.todoapp.theme.AppTheme
-import com.example.todoapp.view.TaskForm
+import com.example.todoapp.ui.component.TaskForm
 import java.util.Date
 
 val task = Task(
