+++ app/src/main/java/com/example/todoapp/ui/screen/EditTaskFormScreen.kt
@@ -1,10 +1,11 @@
-package com.example.todoapp.view
+package com.example.todoapp.ui.screen
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.hilt.navigation.compose.hiltViewModel
-import com.example.todoapp.model.TaskDomain
-import com.example.todoapp.model.toEntity
+import com.example.todoapp.data.repository.model.TaskDomain
+import com.example.todoapp.ui.component.TaskForm
+import com.example.todoapp.utility.toEntity
 import com.example.todoapp.viewmodel.EditTaskFormViewModel
 
 @Composable
