+++ app/src/main/java/com/example/todoapp/viewmodel/EditTaskFormViewModel.kt
@@ -3,8 +3,8 @@ package com.example.todoapp.viewmodel
 import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.todoapp.model.Task
-import com.example.todoapp.usecase.UpdateTaskUseCase
+import com.example.todoapp.data.repository.model.Task
+import com.example.todoapp.domain.usecase.UpdateTaskUseCase
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
 import javax.inject.Inject
